Added new working directory + dataset cleaning
